+++ package.json
@@ -35,6 +35,7 @@
     "jsonwebtoken": "^7.3.0",
     "mongoose": "^4.10.5",
     "nightmare": "^2.10.0",
+    "nodemailer": "^4.0.1",
     "normalize.css": "^7.0.0",
     "pretty-error": "^2.1.0",
     "prop-types": "^15.5.8",
