+++ src/components/Main/index.js
@@ -6,6 +6,7 @@ import DiscreteVector from 'discrete-vector';
 import { Card, CardBlock, Row, Col } from 'reactstrap';
 import withStyles from 'isomorphic-style-loader/lib/withStyles';
 
+import Reports from '../Reports';
 import styles from './styles.scss';
 import Contacts from '../Contacts';
 import LineChart from '../LineChart';
@@ -146,6 +147,7 @@ class Main extends Component {
         moneyRate,
         estimateOptions,
       },
+      headerState,
       addClientData,
       changeMoneyRate,
       addEstimateOptions,
@@ -187,15 +189,20 @@ class Main extends Component {
         </Col>
         <Col xs="12">
           <FinalEstimate
-            mainState={this.props.mainState}
+            moneyRate={moneyRate}
             totalHours={this.state.totalHours}
-            headerState={this.props.headerState}
           />
         </Col>
         <Col xs="12">
           <Contacts
             contacts={contacts}
             addClientData={addClientData}
+          />
+        </Col>
+        <Col xs="12">
+          <Reports
+            headerState={headerState}
+            mainState={this.props.mainState}
           />
         </Col>
       </Row>
